Update more actions
