+++ ROADMAP.md
@@ -1,0 +1,11 @@
+# Roadmap
+
+WIP: This file defines roadmap for project
+
+- [ ]: Custom build system
+- [ ]: Fix pydeps caching timeout
+- [ ]: Make binaries work on any linux distro
+- [ ]: Mac binaries for pyflow and python
+- [ ]: "fatal: destination path exists" when using git deps
+- [ ]: add hash and git/path info to locks
+- [ ]: clear download git source as an option. In general, git install is a mess
